notifications/django-admin-panel/requirements.txt
asgiref==3.5.0
Django==4.0
django-debug-toolbar==3.2.4
django-split-settings==1.1.0
djangorestframework==3.13.1
flake8==4.0.0
gunicorn==20.1.0
mccabe==0.6.1
model-bakery==1.5.0
psycopg2-binary==2.9.3 
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
django-celery-beat==2.3.0
django-celery-results==2.4.0
celery==5.2.7
redis==4.3.4
requests==2.28.1


